Fixing CommonModule bindings for EventListeners
